wildfly-ai-assist/pyproject.toml
[tool.poetry]
name = "wildfly-ai-assist"
version = "0.1.0"
description = "Stack‑trace→wiki Retrieval‑Augmented Generation agent for WildFly"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.30"}
typer = "^0.12"
llama-index-core = "^0.12.45"
llama-index-embeddings-openai = "^0.1.14"
llama-index-llms-openai = "^0.1.12"
openai = "^1.14"
faiss-cpu = "^1.8"
python-dotenv = "^1.0"
atlassian-python-api = "^4.2"
python-slugify = "^8.0"
pytest = "^8.1"
pytest-asyncio = "^0.23"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
isort = "^5.13"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
